node refactoring
