+++ node.go
@@ -4,13 +4,9 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"math"
 	"net"
 	"os"
 	"path/filepath"
-	"sort"
-	"strconv"
-	"strings"
 	"sync"
 	"sync/atomic"
 	"time"
@@ -630,700 +626,6 @@ func (n *Node) electAndStart(bgCtx context.Context) error {
 	return nil
 }
 
-// becomeLeader transitions this node to leader role.
-// Re-opens the WAL with an S3 uploader, starts the peer gRPC server,
-// and launches the watchLoop. Must NOT be called with n.mu held.
-func (n *Node) becomeLeader(bgCtx context.Context, lock *election.Lock, rec *election.LockRecord) error {
-	n.wal.Close()
-	walDir := filepath.Join(n.cfg.DataDir, "wal")
-
-	// Upload any local WAL segments that were not yet in S3 before taking on
-	// writes. This covers the same-node re-election case where the previous WAL
-	// had no uploader (follower WAL) or crashed before the upload completed.
-	// After this point, new leader writes are uploaded async (SegmentMaxAge).
-	upCtx, upCancel := context.WithTimeout(context.Background(), 2*time.Minute)
-	uploadLocalWALSegments(upCtx, walDir, n.cfg.ObjectStore, n.log)
-	upCancel()
-
-	// Replay any remote WAL entries not yet in our Pebble. A follower that wins
-	// election may be behind the former leader if the former leader committed
-	// entries during single-node mode (no quorum required) before crashing.
-	// Check against the latest S3 checkpoint first: if this node is behind the
-	// checkpoint, restore it before replaying WAL so replayRemote only needs
-	// segments still present in S3.
-	if n.cfg.ObjectStore != nil {
-		cpCtx, cpCancel := context.WithTimeout(context.Background(), 5*time.Minute)
-		if _, cpErr := n.restoreDBIfBehindCheckpoint(cpCtx); cpErr != nil {
-			cpCancel()
-			return fmt.Errorf("t4: leader checkpoint catch-up: %w", cpErr)
-		}
-		cpCancel()
-
-		reCtx, reCancel := context.WithTimeout(context.Background(), 2*time.Minute)
-		if err := replayRemote(reCtx, n.db.Load(), n.cfg.ObjectStore, n.db.Load().CurrentRevision(), n.log); err != nil {
-			reCancel()
-			return fmt.Errorf("t4: becomeLeader replay remote WAL: %w", err)
-		}
-		reCancel()
-	}
-
-	// With quorum commit, every committed entry exists on at least two nodes'
-	// WALs before the caller sees success. S3 is disaster-recovery only (both
-	// nodes fail simultaneously), so uploads can be async — driven by
-	// SegmentMaxAge — without affecting write durability.
-	w2, err := wal.Open(walDir, rec.Term, n.db.Load().CurrentRevision()+1,
-		wal.WithUploader(makeUploader(n.cfg.ObjectStore, n.log)),
-		wal.WithSegmentMaxSize(n.cfg.SegmentMaxSize),
-		wal.WithSegmentMaxAge(n.cfg.SegmentMaxAge),
-		wal.WithLogger(n.log),
-	)
-	if err != nil {
-		return fmt.Errorf("t4: open WAL as leader: %w", err)
-	}
-	w2.Start(bgCtx)
-
-	peerSrv := peer.NewServer(n.cfg.PeerBufferSize, n.log)
-	lis, err := net.Listen("tcp", n.cfg.PeerListenAddr)
-	if err != nil {
-		w2.Close()
-		return fmt.Errorf("t4: peer listen %s: %w", n.cfg.PeerListenAddr, err)
-	}
-	serverOpts := []grpc.ServerOption{grpc.ForceServerCodec(peer.Codec{})}
-	if n.cfg.PeerServerTLS != nil {
-		serverOpts = append(serverOpts, grpc.Creds(n.cfg.PeerServerTLS))
-	}
-	grpcSrv := grpc.NewServer(serverOpts...)
-	peer.RegisterWalStreamServer(grpcSrv, peerSrv)
-
-	// Commit state transition atomically before accepting connections.
-	n.mu.Lock()
-	n.wal = w2
-	n.term = rec.Term
-	n.peerSrv = peerSrv
-	n.peerLis = lis
-	n.peerGRPC = grpcSrv
-	n.leaderCli.Store(nil) // leader does not forward writes
-	n.storeRole(roleLeader)
-	n.nextRev = n.db.Load().CurrentRevision() // sync revision counter after any replay
-	n.pending = make(map[string]pendingKV)
-	n.mu.Unlock()
-
-	// Install the forward handler after role is set to leader so that
-	// HandleForward sees the correct role and executes writes directly.
-	peerSrv.SetForwardHandler(n)
-	// Tell the peer server what the first revision this leader will write is.
-	// Followers connecting with a lower fromRev are missing entries that were
-	// only replayed into Pebble from S3 (never in the ring buffer) and must
-	// re-sync before consuming the live stream.
-	peerSrv.SetStartRev(n.db.Load().CurrentRevision() + 1)
-
-	go func() {
-		if err := grpcSrv.Serve(lis); err != nil {
-			n.log.Warnf("t4: peer server: %v", err)
-		}
-	}()
-
-	n.updateMetrics()
-	metrics.ElectionsTotal.WithLabelValues("won").Inc()
-	n.log.Infof("t4: elected leader (term=%d, peer=%s)", rec.Term, n.cfg.PeerListenAddr)
-	go n.watchLoop(bgCtx, lock, rec.Term)
-	return nil
-}
-
-// watchLoop periodically reads the lock from S3 to detect supersession.
-// Steps down (cancelBg) if the lock's term or owner changes.
-// On clean shutdown, releases the lock.
-//
-// Split-brain prevention strategy:
-//
-//  1. Periodic fallback: LeaderWatchInterval reads S3 to detect supersession.
-//     Does not touch LastSeenNano — followers are healthy while connected.
-//
-//  2. On follower disconnect: immediately fence writes (~50ms), verify still
-//     leader, touch LastSeenNano so the disconnected follower backs off from
-//     TakeOver. Begin polling every peer.FollowerRetryInterval to keep
-//     LastSeenNano fresh and detect any TakeOver that follows. Stop polling
-//     once followers reconnect.
-//
-//  3. TakeOver safety: if a follower fails to reconnect after FollowerMaxRetries
-//     it calls TakeOver. If LastSeenNano is older than LeaderLivenessTTL the
-//     TakeOver proceeds (via atomic conditional PUT). The leader detects the
-//     supersession at its next fencedCheck and steps down cleanly.
-func (n *Node) watchLoop(ctx context.Context, lock *election.Lock, term uint64) {
-	ticker := time.NewTicker(n.cfg.LeaderWatchInterval)
-	defer ticker.Stop()
-
-	// Disconnect channel from the peer server, if we're running in multi-node mode.
-	var disconnectC <-chan struct{}
-	if n.peerSrv != nil {
-		disconnectC = n.peerSrv.DisconnectC
-	}
-
-	// pollC is non-nil while liveness-touch polling is active (nil blocks in select).
-	var (
-		activePollTicker *time.Ticker
-		pollC            <-chan time.Time
-	)
-
-	stopPolling := func() {
-		if activePollTicker != nil {
-			activePollTicker.Stop()
-			activePollTicker = nil
-		}
-		pollC = nil
-	}
-
-	startPolling := func() {
-		if activePollTicker != nil {
-			return // already polling
-		}
-		activePollTicker = time.NewTicker(peer.FollowerRetryInterval)
-		pollC = activePollTicker.C
-	}
-
-	// Always start polling immediately in cluster mode so LastSeenNano is
-	// kept fresh from the very first tick, even before any follower connects.
-	// Without this, a leader with no followers never touches the lock and the
-	// liveness record goes stale after LeaderLivenessTTL, letting any
-	// recovering follower win TakeOver and create a split-brain.
-	if n.peerSrv != nil {
-		startPolling()
-	}
-
-	// fencedCheck fences all leader writes for the duration of one S3 GET (and
-	// optional PUT). Returns false and steps down if the lock has been superseded.
-	// When touch is true and still leader, also writes LastSeenNano to the lock
-	// so disconnected followers see a fresh liveness signal and back off TakeOver.
-	//
-	// The Read and the Touch (when requested) are tied together via a conditional
-	// PUT (If-Match: <etag>): if another node wins the lock between our Read and
-	// our Touch, TouchIfMatch returns ErrPreconditionFailed and we step down
-	// immediately — closing the Read→Touch split-brain race.
-	//
-	// NOTE: fenceMu is released explicitly (not via defer) so that grpcSrv.Stop()
-	// can be called outside the lock.  grpc.Server.Stop waits for in-flight
-	// handlers to finish; those handlers (Put/Create/…) acquire fenceMu.RLock()
-	// themselves, so calling Stop() while holding the write lock would deadlock.
-	fencedCheck := func(reason string, touch bool) bool {
-		n.fenceMu.Lock()
-		rCtx, cancel := context.WithTimeout(ctx, 10*time.Second)
-		rec, etag, err := lock.ReadETag(rCtx)
-		cancel()
-		if err != nil {
-			n.fenceMu.Unlock()
-			n.log.Warnf("t4: leader watch (%s): read lock: %v", reason, err)
-			return true // transient S3 error; keep running
-		}
-		if rec == nil || rec.Term != term || rec.NodeID != n.cfg.NodeID {
-			n.log.Errorf("t4: leader watch (%s): lock superseded (current: %+v) — stepping down", reason, rec)
-			n.cancelBg()
-			n.fenceMu.Unlock()
-			// Stop the peer gRPC server so followers immediately lose their
-			// streams and detect the leadership change.  Without this, followers
-			// keep forwarding ForwardGetRevision to this zombie leader and receive
-			// a stale revision, causing linearizability violations.
-			if grpcSrv := n.peerGRPC; grpcSrv != nil {
-				grpcSrv.Stop()
-			}
-			return false
-		}
-		if touch && n.peerSrv != nil {
-			tCtx, tCancel := context.WithTimeout(ctx, 5*time.Second)
-			err := lock.TouchIfMatch(tCtx, term, n.cfg.AdvertisePeerAddr, etag, n.db.Load().CurrentRevision())
-			tCancel()
-			if errors.Is(err, object.ErrPreconditionFailed) {
-				// Another node wrote the lock between our Read and our Touch —
-				// we have been superseded.  Step down immediately.
-				n.log.Errorf("t4: leader watch (%s): touch precondition failed — lock taken, stepping down", reason)
-				n.cancelBg()
-				n.fenceMu.Unlock()
-				if grpcSrv := n.peerGRPC; grpcSrv != nil {
-					grpcSrv.Stop()
-				}
-				return false
-			}
-			if err != nil {
-				n.log.Warnf("t4: leader watch (%s): touch lock: %v", reason, err)
-			}
-		}
-		n.fenceMu.Unlock()
-		return true
-	}
-
-	for {
-		select {
-		case <-ticker.C:
-			if !fencedCheck("periodic", false) {
-				return
-			}
-
-		case <-disconnectC:
-			// Immediate check + touch: catches TakeOver that already happened,
-			// and signals liveness to the disconnected follower.
-			n.log.Infof("t4: leader watch: follower disconnected — fencing writes and checking lock")
-			if !fencedCheck("disconnect", true) {
-				return
-			}
-			// Resume polling so LastSeenNano stays fresh while the follower
-			// is away.  startPolling is idempotent — if we were already polling
-			// (e.g. started at launch with no followers), this is a no-op.
-			startPolling()
-
-		case <-pollC:
-			// Touch the lock to signal liveness and detect supersession.
-			if !fencedCheck("poll", true) {
-				stopPolling()
-				return
-			}
-			// If followers have (re)connected, liveness touches are no longer
-			// necessary — stop polling and rely on the periodic ticker.
-			if n.peerSrv != nil && n.peerSrv.ConnectedFollowers() > 0 {
-				n.log.Debugf("t4: leader watch: followers connected — pausing liveness poll")
-				stopPolling()
-			}
-
-		case <-ctx.Done():
-			stopPolling()
-			return
-		}
-	}
-}
-
-// followLoop receives WAL entries from the leader and applies them locally.
-// On ErrLeaderUnreachable it attempts a TakeOver election.
-func (n *Node) followLoop(bgCtx context.Context) {
-	lock := election.NewLock(n.cfg.ObjectStore, n.cfg.NodeID, n.cfg.AdvertisePeerAddr)
-	cli := n.peerCli
-	fromRev := n.db.Load().CurrentRevision() + 1
-
-	for {
-		err := cli.Follow(
-			bgCtx,
-			fromRev,
-			func(entries []wal.Entry) error {
-				// Followers must apply a contiguous revision stream. If the leader
-				// stream skips (or rewinds) a revision, force a full resync rather
-				// than silently advancing currentRev with holes.
-				for i, e := range entries {
-					if e.Revision != fromRev+int64(i) {
-						return peer.ErrResyncRequired
-					}
-				}
-				ptrs := make([]*wal.Entry, len(entries))
-				for i := range entries {
-					ptrs[i] = &entries[i]
-				}
-				if err := n.wal.AppendBatch(bgCtx, ptrs); err != nil {
-					return err
-				}
-				return nil
-			},
-			func(entries []wal.Entry) error {
-				if err := n.db.Load().Apply(entries); err != nil {
-					return err
-				}
-				// Track the leader's term so attemptPromotion uses the correct
-				// floorTerm when calling TakeOver.  Without this, n.term stays at
-				// its Open() value and TakeOver backs off because it sees the
-				// current lock term as "already taken over at a higher term".
-				// The last entry in the batch has the highest-or-equal term.
-				if last := entries[len(entries)-1]; last.Term > n.term {
-					n.mu.Lock()
-					if last.Term > n.term {
-						n.term = last.Term
-					}
-					n.mu.Unlock()
-				}
-				// Advance only after a successful apply so a reconnect retries
-				// from the start of the failed batch rather than skipping it.
-				fromRev = entries[len(entries)-1].Revision + 1
-				return nil
-			},
-		)
-
-		if bgCtx.Err() != nil {
-			return
-		}
-
-		if peer.IsResyncRequired(err) {
-			metrics.FollowerResyncsTotal.WithLabelValues("stream_gap").Inc()
-			if n.cfg.ObjectStore == nil {
-				n.log.Errorf("t4: follower resync required but no object store — restart node")
-				n.cancelBg()
-				return
-			}
-			// Ring buffer miss: the follower has been offline long enough that
-			// the leader's ring buffer no longer covers fromRev. Restore from
-			// the latest S3 checkpoint (if the follower's Pebble is behind it),
-			// then replay any remaining WAL entries from S3.
-			n.log.Warnf("t4: follower resync required — restoring from checkpoint")
-			if cpErr := n.resyncFromCheckpoint(bgCtx); cpErr != nil {
-				n.log.Errorf("t4: follower in-place resync failed: %v — cancelling", cpErr)
-				n.cancelBg()
-				return
-			}
-			reCtx, reCancel := context.WithTimeout(bgCtx, 5*time.Minute)
-			rerr := replayRemote(reCtx, n.db.Load(), n.cfg.ObjectStore, n.db.Load().CurrentRevision(), n.log)
-			reCancel()
-			if rerr != nil {
-				n.log.Errorf("t4: follower S3 resync failed: %v — retrying", rerr)
-				select {
-				case <-time.After(2 * time.Second):
-				case <-bgCtx.Done():
-					return
-				}
-			} else {
-				fromRev = n.db.Load().CurrentRevision() + 1
-				// Wake any goroutines blocked in WaitForRevision that entered
-				// their wait loop while replayRemote was running. Recover does
-				// not broadcast, so without this they would sleep until the
-				// next live Apply — causing unnecessary read latency.
-				n.db.Load().NotifyRevision()
-				n.log.Infof("t4: follower resync complete (now at rev=%d)", n.db.Load().CurrentRevision())
-			}
-			continue
-		}
-
-		if peer.IsLeaderUnreachable(err) || peer.IsLeaderShutdown(err) {
-			if peer.IsLeaderShutdown(err) {
-				n.log.Infof("t4: leader shut down gracefully — attempting immediate election takeover")
-			} else {
-				n.log.Warnf("t4: leader unreachable — attempting election takeover")
-			}
-			newCli, promoted := n.attemptPromotion(bgCtx, lock, peer.IsLeaderShutdown(err))
-			if promoted {
-				return
-			}
-			if newCli != nil {
-				oldCli := cli
-				cli = newCli
-				n.leaderCli.Store(newCli)
-				oldCli.Close()
-				n.log.Infof("t4: following new leader")
-			}
-			continue
-		}
-
-		n.log.Warnf("t4: follow loop error (will retry): %v", err)
-		select {
-		case <-time.After(2 * time.Second):
-		case <-bgCtx.Done():
-			return
-		}
-	}
-}
-
-// attemptPromotion tries to take over the leader lock after the stream dies.
-// Returns (nil, true) if promoted to leader.
-// Returns (newClient, false) if another node won; newClient follows that node.
-// Returns (nil, false) on S3 errors or when the current leader's liveness
-// record is fresh enough that TakeOver would risk a split-brain.
-//
-// graceful should be true when the leader sent an explicit shutdown signal:
-// in that case the liveness check is skipped because the leader intentionally
-// vacated and the fresh LastSeenNano would otherwise block all followers.
-func (n *Node) attemptPromotion(bgCtx context.Context, lock *election.Lock, graceful bool) (*peer.Client, bool) {
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
-	// Read the current lock before attempting TakeOver. If the leader recently
-	// wrote a liveness touch (LastSeenNano is fresh) AND the shutdown was not
-	// graceful, the leader may still have other followers and we are an
-	// isolated minority node — back off to prevent split-brain.
-	// On graceful shutdown we skip this check: the leader intentionally left,
-	// so its fresh LastSeenNano must not block election.
-	existing, err := lock.Read(ctx)
-	if err != nil {
-		n.log.Errorf("t4: takeover: read lock: %v", err)
-		return nil, false
-	}
-	if !graceful && existing != nil && existing.LastSeenNano > 0 {
-		age := time.Since(time.Unix(0, existing.LastSeenNano))
-		if age < peer.LeaderLivenessTTL {
-			n.log.Infof("t4: takeover: leader liveness is fresh (%v ago) — backing off to avoid split-brain", age.Round(time.Millisecond))
-			// Back off from election, but do not keep retrying a stale endpoint.
-			// If the lock advertises a leader address, switch followLoop to it.
-			if existing.LeaderAddr != "" {
-				return peer.NewClient(existing.LeaderAddr, n.cfg.NodeID, n.cfg.FollowerMaxRetries, n.cfg.PeerClientTLS, n.log), false
-			}
-			return nil, false
-		}
-	}
-	// Revision fence: refuse to become leader if we are behind the last known
-	// committed revision. A node missing entries would either drop them (data
-	// loss) or fail to serve reads that clients already observed.
-	// If the lock already points to a current leader, return their address so
-	// followLoop switches to following them — connecting will trigger an
-	// in-place resync that catches up this node's revision.
-	if existing != nil && existing.CommittedRev > n.db.Load().CurrentRevision() {
-		n.log.Infof("t4: takeover: node is behind leader committed rev (ours=%d, leader=%d) — following current leader to catch up",
-			n.db.Load().CurrentRevision(), existing.CommittedRev)
-		if existing.LeaderAddr != "" {
-			return peer.NewClient(existing.LeaderAddr, n.cfg.NodeID, n.cfg.FollowerMaxRetries, n.cfg.PeerClientTLS, n.log), false
-		}
-		return nil, false
-	}
-
-	rec, won, err := lock.TakeOver(ctx, n.term, n.db.Load().CurrentRevision())
-	if err != nil {
-		n.log.Errorf("t4: takeover election error: %v", err)
-		return nil, false
-	}
-
-	if won {
-		if err := n.becomeLeader(bgCtx, lock, rec); err != nil {
-			n.log.Errorf("t4: promotion failed: %v", err)
-			return nil, false
-		}
-		// Start write-processing loops immediately so that client writes are
-		// not blocked while we run Reconcile and the startup checkpoint below.
-		// forceCheckpoint (and periodic maybeCheckpoint) already hold
-		// fenceMu.Lock() during I/O, which briefly pauses new writes — that
-		// is the same behaviour as during a normal checkpoint interval.
-		n.bgWg.Add(1)
-		go func() { defer n.bgWg.Done(); n.commitLoop(bgCtx) }()
-		if n.cfg.ObjectStore != nil && n.cfg.CheckpointInterval > 0 {
-			n.bgWg.Add(1)
-			go func() { defer n.bgWg.Done(); n.checkpointLoop(bgCtx) }()
-		}
-		// Upload any SSTs that exist on disk but aren't in S3 yet. The
-		// follower didn't run SSTUploader.Start(), so its SSTs were never
-		// streamed. Additionally, becomeLeader may have restored from a
-		// checkpoint and replayed WAL, creating new SST files. Reconcile
-		// ensures all of them are in S3 before the first checkpoint.
-		if n.sstUploader != nil {
-			rCtx, rCancel := context.WithTimeout(context.Background(), 2*time.Minute)
-			if rErr := n.sstUploader.Reconcile(rCtx); rErr != nil {
-				n.log.Warnf("t4: promoted leader SST reconcile: %v", rErr)
-			}
-			rCancel()
-			n.sstUploader.Start(bgCtx)
-		}
-		// Write a checkpoint immediately after Reconcile so that all
-		// uploaded SSTs are referenced by a live checkpoint. Without this,
-		// the old leader's GCOrphanSSTs could delete the just-uploaded SSTs
-		// before the checkpointLoop gets a chance to write its startup
-		// checkpoint. The checkpointLoop's own forceCheckpoint is redundant
-		// but harmless (same rev → same checkpoint key → idempotent overwrite).
-		if n.cfg.ObjectStore != nil && n.cfg.CheckpointInterval > 0 {
-			n.forceCheckpoint(bgCtx)
-		}
-		return nil, true
-	}
-
-	if rec != nil && rec.LeaderAddr != "" {
-		n.log.Infof("t4: lost election to %s (term=%d) — following", rec.NodeID, rec.Term)
-		return peer.NewClient(rec.LeaderAddr, n.cfg.NodeID, n.cfg.FollowerMaxRetries, n.cfg.PeerClientTLS, n.log), false
-	}
-	return nil, false
-}
-
-// ── Write forwarding ──────────────────────────────────────────────────────────
-
-// HandleForward implements peer.ForwardHandler. Called by the peer gRPC server
-// when a follower forwards a write. Dispatches to the appropriate Node method.
-// Since HandleForward runs on the leader, all write methods execute directly.
-func (n *Node) HandleForward(ctx context.Context, req *peer.ForwardRequest) (*peer.ForwardResponse, error) {
-	switch req.Op {
-	case peer.ForwardPut:
-		rev, err := n.Put(ctx, req.Key, req.Value, req.Lease)
-		code, msg := encodeErr(err)
-		return &peer.ForwardResponse{Revision: rev, Succeeded: err == nil, ErrCode: code, ErrMsg: msg}, nil
-
-	case peer.ForwardCreate:
-		rev, err := n.Create(ctx, req.Key, req.Value, req.Lease)
-		code, msg := encodeErr(err)
-		return &peer.ForwardResponse{Revision: rev, Succeeded: err == nil, ErrCode: code, ErrMsg: msg}, nil
-
-	case peer.ForwardUpdate:
-		newRev, oldKV, updated, err := n.Update(ctx, req.Key, req.Value, req.Revision, req.Lease)
-		code, msg := encodeErr(err)
-		resp := &peer.ForwardResponse{Revision: newRev, Succeeded: updated, ErrCode: code, ErrMsg: msg}
-		resp.OldKV = kvToMsg(oldKV)
-		return resp, nil
-
-	case peer.ForwardDeleteIfRevision:
-		newRev, oldKV, deleted, err := n.DeleteIfRevision(ctx, req.Key, req.Revision)
-		code, msg := encodeErr(err)
-		resp := &peer.ForwardResponse{Revision: newRev, Succeeded: deleted, ErrCode: code, ErrMsg: msg}
-		resp.OldKV = kvToMsg(oldKV)
-		return resp, nil
-
-	case peer.ForwardCompact:
-		err := n.Compact(ctx, req.Revision)
-		code, msg := encodeErr(err)
-		return &peer.ForwardResponse{Succeeded: err == nil, ErrCode: code, ErrMsg: msg}, nil
-
-	case peer.ForwardGetRevision:
-		// Return nextRev (the highest *assigned* revision), not db.CurrentRevision()
-		// (the last *applied* revision). A write increments nextRev under n.mu and
-		// sends to writeC before the commit loop applies it to Pebble. If we returned
-		// db.CurrentRevision() here, a follower could sync to a revision that precedes
-		// an in-flight write whose acknowledgment is about to be sent to the client —
-		// causing a stale read that violates linearizability.
-		n.mu.Lock()
-		rev := n.nextRev
-		n.mu.Unlock()
-		return &peer.ForwardResponse{Revision: rev, Succeeded: true}, nil
-
-	case peer.ForwardTxn:
-		if req.TxnReq == nil {
-			return nil, fmt.Errorf("t4: ForwardTxn missing TxnReq")
-		}
-		txnReq := forwardMsgToTxnRequest(req.TxnReq)
-		resp, err := n.Txn(ctx, txnReq)
-		if err != nil {
-			code, msg := encodeErr(err)
-			return &peer.ForwardResponse{ErrCode: code, ErrMsg: msg}, nil
-		}
-		deletedKeys := make([]string, 0, len(resp.DeletedKeys))
-		for k := range resp.DeletedKeys {
-			deletedKeys = append(deletedKeys, k)
-		}
-		return &peer.ForwardResponse{
-			Revision:    resp.Revision,
-			Succeeded:   resp.Succeeded,
-			DeletedKeys: deletedKeys,
-		}, nil
-	}
-	return nil, fmt.Errorf("t4: unknown forward op %d", req.Op)
-}
-
-// forwardWrite sends a write request to the leader and decodes the response.
-func (n *Node) forwardWrite(ctx context.Context, req *peer.ForwardRequest) (*peer.ForwardResponse, error) {
-	cli := n.leaderCli.Load()
-	if cli == nil {
-		return nil, ErrNotLeader
-	}
-	op := fwdOpLabel(req.Op)
-	start := time.Now()
-	resp, err := cli.ForwardWrite(ctx, req)
-	metrics.ForwardedWritesTotal.WithLabelValues(op).Inc()
-	metrics.ForwardDuration.WithLabelValues(op).Observe(time.Since(start).Seconds())
-	return resp, err
-}
-
-func fwdOpLabel(op peer.ForwardOp) string {
-	switch op {
-	case peer.ForwardPut:
-		return "put"
-	case peer.ForwardCreate:
-		return "create"
-	case peer.ForwardUpdate:
-		return "update"
-	case peer.ForwardDeleteIfRevision:
-		return "delete"
-	case peer.ForwardCompact:
-		return "compact"
-	case peer.ForwardGetRevision:
-		return "get_revision"
-	case peer.ForwardTxn:
-		return "txn"
-	default:
-		return "unknown"
-	}
-}
-
-func encodeErr(err error) (code, msg string) {
-	if err == nil {
-		return "", ""
-	}
-	if errors.Is(err, ErrKeyExists) {
-		return "key_exists", ""
-	}
-	return "error", err.Error()
-}
-
-func decodeErr(code, msg string) error {
-	switch code {
-	case "":
-		return nil
-	case "key_exists":
-		return ErrKeyExists
-	default:
-		return errors.New(msg)
-	}
-}
-
-func kvToMsg(kv *KeyValue) *peer.KVMsg {
-	if kv == nil {
-		return nil
-	}
-	return &peer.KVMsg{
-		Key: kv.Key, Value: kv.Value, Revision: kv.Revision,
-		CreateRevision: kv.CreateRevision, PrevRevision: kv.PrevRevision,
-		Lease: kv.Lease,
-	}
-}
-
-func msgToKV(m *peer.KVMsg) *KeyValue {
-	if m == nil {
-		return nil
-	}
-	return &KeyValue{
-		Key: m.Key, Value: m.Value, Revision: m.Revision,
-		CreateRevision: m.CreateRevision, PrevRevision: m.PrevRevision,
-		Lease: m.Lease,
-	}
-}
-
-// ReadConsistency returns the configured read consistency mode.
-func (n *Node) ReadConsistency() ReadConsistency { return n.cfg.ReadConsistency }
-
-// syncWithLeader implements the ReadIndex pattern: ask the leader for its
-// current revision, then wait until this node has applied at least that far.
-// Returns nil immediately if the node is the leader or running single-node.
-func (n *Node) syncWithLeader(ctx context.Context) error {
-	cli := n.leaderCli.Load()
-	if cli == nil {
-		// If the background context has been cancelled the node is either
-		// shutting down or has been fenced (leader superseded by a new term).
-		// Serving a read from our local stale Pebble would violate
-		// linearizability — return an error so the client retries elsewhere.
-		if n.bgCtx.Err() != nil {
-			return ErrClosed
-		}
-		return nil // leader or single-node — already up-to-date
-	}
-	resp, err := cli.ForwardWrite(ctx, &peer.ForwardRequest{Op: peer.ForwardGetRevision})
-	if err != nil {
-		return fmt.Errorf("t4: read sync: %w", err)
-	}
-	if err := n.WaitForRevision(ctx, resp.Revision); err != nil {
-		if errors.Is(err, context.DeadlineExceeded) || errors.Is(err, context.Canceled) {
-			return fmt.Errorf("t4: read sync: leader reported rev=%d but local node only reached rev=%d before wait ended: %w",
-				resp.Revision, n.db.Load().CurrentRevision(), err)
-		}
-		return fmt.Errorf("t4: read sync: wait for local revision %d: %w", resp.Revision, err)
-	}
-	return nil
-}
-
-// LinearizableGet returns the value for key with linearizability guaranteed.
-// On a follower it syncs to the leader's revision before serving locally.
-func (n *Node) LinearizableGet(ctx context.Context, key string) (*KeyValue, error) {
-	if err := n.syncWithLeader(ctx); err != nil {
-		return nil, err
-	}
-	return n.Get(key)
-}
-
-// LinearizableList returns all keys with the given prefix with linearizability guaranteed.
-func (n *Node) LinearizableList(ctx context.Context, prefix string) ([]*KeyValue, error) {
-	if err := n.syncWithLeader(ctx); err != nil {
-		return nil, err
-	}
-	return n.List(prefix)
-}
-
-// LinearizableCount returns the count of keys with the given prefix with linearizability guaranteed.
-func (n *Node) LinearizableCount(ctx context.Context, prefix string) (int64, error) {
-	if err := n.syncWithLeader(ctx); err != nil {
-		return 0, err
-	}
-	return n.Count(prefix)
-}
-
 // Close shuts down the node cleanly.
 func (n *Node) Close() error {
 	var err error
@@ -1411,875 +713,64 @@ func (n *Node) Close() error {
 	return err
 }
 
-// ── Write path (leader / single-node execute; follower forwards) ──────────────
-
-// Put creates or updates key with value. Returns the new revision.
-func (n *Node) Put(ctx context.Context, key string, value []byte, lease int64) (int64, error) {
-	if n.closed.Load() {
-		return 0, ErrClosed
-	}
-	if n.loadRole() == roleFollower {
-		resp, err := n.forwardWrite(ctx, &peer.ForwardRequest{Op: peer.ForwardPut, Key: key, Value: value, Lease: lease})
-		if err != nil {
-			return 0, err
-		}
-		return resp.Revision, decodeErr(resp.ErrCode, resp.ErrMsg)
-	}
-	n.fenceMu.RLock()
-	defer n.fenceMu.RUnlock()
-	start := time.Now()
-	n.mu.Lock()
-	if n.closed.Load() {
-		n.mu.Unlock()
-		return 0, ErrClosed
-	}
-	e, err := n.preparePut(key, value, lease)
-	if err != nil {
-		n.mu.Unlock()
-		return 0, err
-	}
-	req := newWriteReq(ctx, e)
-	n.writeC <- req
-	n.mu.Unlock()
-	return n.await(ctx, req, opLabel(e.Op), start, key, e.Revision)
-}
-
-// preparePut builds the WAL entry for a Put. Must be called under n.mu.
-func (n *Node) preparePut(key string, value []byte, lease int64) (wal.Entry, error) {
-	existing, err := n.readKey(key)
-	if err != nil {
-		return wal.Entry{}, err
-	}
-	n.nextRev++
-	newRev := n.nextRev
-	var op wal.Op
-	var createRev, prevRev int64
-	if existing == nil {
-		op, createRev = wal.OpCreate, newRev
-	} else {
-		op, createRev, prevRev = wal.OpUpdate, existing.CreateRevision, existing.Revision
-	}
-	n.pending[key] = pendingKV{
-		rev: newRev,
-		kv: &istore.KeyValue{
-			Key: key, Value: value, Revision: newRev,
-			CreateRevision: createRev, PrevRevision: prevRev,
-			Lease: lease,
-		},
-	}
-	return wal.Entry{
-		Revision: newRev, Term: n.term, Op: op,
-		Key: key, Value: value, Lease: lease,
-		CreateRevision: createRev, PrevRevision: prevRev,
-	}, nil
-}
-
-// Create creates key only if it does not already exist.
-func (n *Node) Create(ctx context.Context, key string, value []byte, lease int64) (int64, error) {
-	if n.loadRole() == roleFollower {
-		resp, err := n.forwardWrite(ctx, &peer.ForwardRequest{Op: peer.ForwardCreate, Key: key, Value: value, Lease: lease})
-		if err != nil {
-			return 0, err
-		}
-		return resp.Revision, decodeErr(resp.ErrCode, resp.ErrMsg)
-	}
-	n.fenceMu.RLock()
-	defer n.fenceMu.RUnlock()
-	start := time.Now()
-	n.mu.Lock()
-	if n.closed.Load() {
-		n.mu.Unlock()
-		return 0, ErrClosed
-	}
-	existing, err := n.readKey(key)
-	if err != nil {
-		n.mu.Unlock()
-		return 0, err
-	}
-	if existing != nil {
-		n.mu.Unlock()
-		return 0, ErrKeyExists
-	}
-	n.nextRev++
-	newRev := n.nextRev
-	n.pending[key] = pendingKV{
-		rev: newRev,
-		kv: &istore.KeyValue{
-			Key: key, Value: value, Revision: newRev,
-			CreateRevision: newRev, Lease: lease,
-		},
-	}
-	e := wal.Entry{
-		Revision: newRev, Term: n.term, Op: wal.OpCreate,
-		Key: key, Value: value, Lease: lease, CreateRevision: newRev,
-	}
-	req := newWriteReq(ctx, e)
-	n.writeC <- req
-	n.mu.Unlock()
-	return n.await(ctx, req, "create", start, key, newRev)
-}
-
-// Update updates key only if its current revision matches (CAS).
-func (n *Node) Update(ctx context.Context, key string, value []byte, revision, lease int64) (int64, *KeyValue, bool, error) {
-	if n.loadRole() == roleFollower {
-		resp, err := n.forwardWrite(ctx, &peer.ForwardRequest{Op: peer.ForwardUpdate, Key: key, Value: value, Revision: revision, Lease: lease})
-		if err != nil {
-			return 0, nil, false, err
-		}
-		return resp.Revision, msgToKV(resp.OldKV), resp.Succeeded, decodeErr(resp.ErrCode, resp.ErrMsg)
-	}
-	n.fenceMu.RLock()
-	defer n.fenceMu.RUnlock()
-	start := time.Now()
-	n.mu.Lock()
-	if n.closed.Load() {
-		n.mu.Unlock()
-		return 0, nil, false, ErrClosed
-	}
-	existing, err := n.readKey(key)
-	if err != nil {
-		n.mu.Unlock()
-		return 0, nil, false, err
-	}
-	if existing == nil || existing.Revision != revision {
-		curRev := n.db.Load().CurrentRevision()
-		n.mu.Unlock()
-		return curRev, toKV(existing), false, nil
-	}
-	n.nextRev++
-	newRev := n.nextRev
-	n.pending[key] = pendingKV{
-		rev: newRev,
-		kv: &istore.KeyValue{
-			Key: key, Value: value, Revision: newRev,
-			CreateRevision: existing.CreateRevision, PrevRevision: existing.Revision,
-			Lease: lease,
-		},
-	}
-	e := wal.Entry{
-		Revision: newRev, Term: n.term, Op: wal.OpUpdate,
-		Key: key, Value: value, Lease: lease,
-		CreateRevision: existing.CreateRevision, PrevRevision: existing.Revision,
-	}
-	oldKV := toKV(existing)
-	req := newWriteReq(ctx, e)
-	n.writeC <- req
-	n.mu.Unlock()
-	newRev, err = n.await(ctx, req, "update", start, key, newRev)
-	if err != nil {
-		return 0, nil, false, err
-	}
-	return newRev, oldKV, true, nil
-}
-
-// Delete removes key unconditionally.
-func (n *Node) Delete(ctx context.Context, key string) (int64, error) {
-	if n.loadRole() == roleFollower {
-		resp, err := n.forwardWrite(ctx, &peer.ForwardRequest{Op: peer.ForwardDeleteIfRevision, Key: key, Revision: 0})
-		if err != nil {
-			return 0, err
-		}
-		return resp.Revision, decodeErr(resp.ErrCode, resp.ErrMsg)
-	}
-	n.fenceMu.RLock()
-	defer n.fenceMu.RUnlock()
-	start := time.Now()
-	n.mu.Lock()
-	if n.closed.Load() {
-		n.mu.Unlock()
-		return 0, ErrClosed
-	}
-	e, err := n.prepareDelete(key)
-	if err != nil || e.Key == "" {
-		n.mu.Unlock()
-		return 0, err // key not found — no-op
-	}
-	req := newWriteReq(ctx, e)
-	n.writeC <- req
-	n.mu.Unlock()
-	return n.await(ctx, req, "delete", start, key, e.Revision)
-}
-
-// DeleteIfRevision deletes key only if its current revision matches (CAS).
-func (n *Node) DeleteIfRevision(ctx context.Context, key string, revision int64) (int64, *KeyValue, bool, error) {
-	if n.loadRole() == roleFollower {
-		resp, err := n.forwardWrite(ctx, &peer.ForwardRequest{Op: peer.ForwardDeleteIfRevision, Key: key, Revision: revision})
-		if err != nil {
-			return 0, nil, false, err
-		}
-		return resp.Revision, msgToKV(resp.OldKV), resp.Succeeded, decodeErr(resp.ErrCode, resp.ErrMsg)
-	}
-	n.fenceMu.RLock()
-	defer n.fenceMu.RUnlock()
-	start := time.Now()
-	n.mu.Lock()
-	if n.closed.Load() {
-		n.mu.Unlock()
-		return 0, nil, false, ErrClosed
-	}
-	existing, err := n.readKey(key)
-	if err != nil {
-		n.mu.Unlock()
-		return 0, nil, false, err
-	}
-	curRev := n.db.Load().CurrentRevision()
-	if existing == nil {
-		n.mu.Unlock()
-		return curRev, nil, false, nil
-	}
-	if revision != 0 && existing.Revision != revision {
-		n.mu.Unlock()
-		return curRev, toKV(existing), false, nil
-	}
-	oldKV := toKV(existing)
-	e, err := n.prepareDelete(key)
-	if err != nil || e.Key == "" {
-		n.mu.Unlock()
-		return 0, nil, false, err
-	}
-	req := newWriteReq(ctx, e)
-	n.writeC <- req
-	n.mu.Unlock()
-	newRev, err := n.await(ctx, req, "delete", start, key, e.Revision)
-	if err != nil {
-		return 0, nil, false, err
-	}
-	return newRev, oldKV, true, nil
-}
-
-// prepareDelete builds a WAL delete entry for key. Must be called under n.mu.
-// Returns a zero-valued entry (Key=="") if the key does not exist.
-func (n *Node) prepareDelete(key string) (wal.Entry, error) {
-	existing, err := n.readKey(key)
-	if err != nil || existing == nil {
-		return wal.Entry{}, err
-	}
-	n.nextRev++
-	newRev := n.nextRev
-	n.pending[key] = pendingKV{rev: newRev, deleted: true}
-	return wal.Entry{
-		Revision: newRev, Term: n.term, Op: wal.OpDelete,
-		Key: key, CreateRevision: existing.CreateRevision, PrevRevision: existing.Revision,
-	}, nil
-}
-
-// readKey returns the current value of key, checking in-flight pending writes
-// first, then falling back to Pebble. Must be called under n.mu.
-func (n *Node) readKey(key string) (*istore.KeyValue, error) {
-	if p, ok := n.pending[key]; ok {
-		if p.deleted {
-			return nil, nil
-		}
-		// pending is an optimistic local read path: once a write has been
-		// assigned a revision under n.mu, concurrent writes to the same key
-		// observe that in-flight state before the batch is durably applied.
-		return p.kv, nil
-	}
-	return n.db.Load().Get(key)
-}
-
-// txnCondMatches reports whether existing satisfies cond.
-// existing is nil when the key does not exist.
-//
-// Note on TxnCondVersion: t4 does not track a per-key write count, so Version
-// is treated as a binary presence flag — 0 means the key is absent, 1 means it
-// is present. TxnCondGreater and TxnCondLess on Version are therefore
-// semantically equivalent to "> 0" and "< 1" respectively; callers that need
-// accurate multi-version comparisons should use TxnCondMod instead.
-func txnCondMatches(cond TxnCondition, existing *istore.KeyValue) bool {
-	var (
-		modRev    int64
-		createRev int64
-		version   int64
-		lease     int64
-		val       []byte
-	)
-	if existing != nil {
-		modRev = existing.Revision
-		createRev = existing.CreateRevision
-		version = 1 // t4 does not track per-key write count; treat as "exists"
-		lease = existing.Lease
-		val = existing.Value
-	}
-
-	var lhs, rhs int64
-	switch cond.Target {
-	case TxnCondMod:
-		lhs, rhs = modRev, cond.ModRevision
-	case TxnCondVersion:
-		lhs, rhs = version, cond.Version
-	case TxnCondCreate:
-		lhs, rhs = createRev, cond.CreateRevision
-	case TxnCondLease:
-		lhs, rhs = lease, cond.Lease
-	case TxnCondValue:
-		switch cond.Result {
-		case TxnCondEqual:
-			return string(val) == string(cond.Value)
-		case TxnCondNotEqual:
-			return string(val) != string(cond.Value)
-		default:
-			return false // ordering on Value bytes not supported
-		}
-	default:
-		return false
-	}
-	switch cond.Result {
-	case TxnCondEqual:
-		return lhs == rhs
-	case TxnCondNotEqual:
-		return lhs != rhs
-	case TxnCondGreater:
-		return lhs > rhs
-	case TxnCondLess:
-		return lhs < rhs
-	default:
-		return false
-	}
-}
-
-// Txn executes an atomic multi-key transaction.
-//
-// All Conditions are evaluated under the write lock in a single atomic step.
-// If every condition is satisfied the Success ops are applied and Succeeded is
-// true; otherwise the Failure ops are applied and Succeeded is false.
-// Either branch may be empty — if no ops need to be written the method returns
-// immediately with the current revision.
-//
-// All write ops within the selected branch share a single revision and are
-// committed to the WAL in one entry, ensuring crash-safe atomicity.
-func (n *Node) Txn(ctx context.Context, req TxnRequest) (TxnResponse, error) {
-	if n.closed.Load() {
-		return TxnResponse{}, ErrClosed
-	}
-	if n.loadRole() == roleFollower {
-		resp, err := n.forwardWrite(ctx, txnToForwardRequest(req))
-		if err != nil {
-			return TxnResponse{}, err
-		}
-		deletedKeys := make(map[string]struct{}, len(resp.DeletedKeys))
-		for _, k := range resp.DeletedKeys {
-			deletedKeys[k] = struct{}{}
-		}
-		return TxnResponse{
-			Succeeded:   resp.Succeeded,
-			Revision:    resp.Revision,
-			DeletedKeys: deletedKeys,
-		}, decodeErr(resp.ErrCode, resp.ErrMsg)
-	}
-	n.fenceMu.RLock()
-	defer n.fenceMu.RUnlock()
-	start := time.Now()
-	n.mu.Lock()
-	if n.closed.Load() {
-		n.mu.Unlock()
-		return TxnResponse{}, ErrClosed
-	}
-	e, succeeded, deletedKeys, err := n.prepareTxn(req)
-	if err != nil {
-		n.mu.Unlock()
-		return TxnResponse{}, err
-	}
-	if e.Op == 0 {
-		// No-op branch: conditions evaluated but no writes needed.
-		// Use the committed revision, not nextRev which may be ahead of what
-		// the commit loop has fsynced if concurrent writes are in flight.
-		curRev := n.db.Load().CurrentRevision()
-		n.mu.Unlock()
-		return TxnResponse{Succeeded: succeeded, Revision: curRev}, nil
-	}
-	wr := newWriteReq(ctx, e)
-	n.writeC <- wr
-	n.mu.Unlock()
-	rev, err := n.await(ctx, wr, "txn", start, "", e.Revision)
-	if err != nil {
-		return TxnResponse{}, err
-	}
-	return TxnResponse{Succeeded: succeeded, Revision: rev, DeletedKeys: deletedKeys}, nil
-}
-
-// txnToForwardRequest converts a TxnRequest into a peer ForwardRequest for
-// follower-to-leader forwarding.
-func txnToForwardRequest(req TxnRequest) *peer.ForwardRequest {
-	conds := make([]peer.TxnCondMsg, len(req.Conditions))
-	for i, c := range req.Conditions {
-		conds[i] = peer.TxnCondMsg{
-			Key:            c.Key,
-			Target:         uint8(c.Target),
-			Result:         uint8(c.Result),
-			ModRevision:    c.ModRevision,
-			CreateRevision: c.CreateRevision,
-			Version:        c.Version,
-			Value:          c.Value,
-			Lease:          c.Lease,
-		}
-	}
-	success := txnOpsToMsg(req.Success)
-	failure := txnOpsToMsg(req.Failure)
-	return &peer.ForwardRequest{
-		Op: peer.ForwardTxn,
-		TxnReq: &peer.TxnReqMsg{
-			Conditions: conds,
-			Success:    success,
-			Failure:    failure,
-		},
-	}
-}
-
-func txnOpsToMsg(ops []TxnOp) []peer.TxnOpMsg {
-	if len(ops) == 0 {
-		return nil
-	}
-	msgs := make([]peer.TxnOpMsg, len(ops))
-	for i, op := range ops {
-		msgs[i] = peer.TxnOpMsg{
-			Type:  uint8(op.Type),
-			Key:   op.Key,
-			Value: op.Value,
-			Lease: op.Lease,
-		}
-	}
-	return msgs
-}
-
-// forwardMsgToTxnRequest converts a peer TxnReqMsg back to a TxnRequest.
-func forwardMsgToTxnRequest(m *peer.TxnReqMsg) TxnRequest {
-	conds := make([]TxnCondition, len(m.Conditions))
-	for i, c := range m.Conditions {
-		conds[i] = TxnCondition{
-			Key:            c.Key,
-			Target:         TxnCondTarget(c.Target),
-			Result:         TxnCondResult(c.Result),
-			ModRevision:    c.ModRevision,
-			CreateRevision: c.CreateRevision,
-			Version:        c.Version,
-			Value:          c.Value,
-			Lease:          c.Lease,
-		}
-	}
-	return TxnRequest{
-		Conditions: conds,
-		Success:    msgToTxnOps(m.Success),
-		Failure:    msgToTxnOps(m.Failure),
-	}
-}
-
-func msgToTxnOps(msgs []peer.TxnOpMsg) []TxnOp {
-	if len(msgs) == 0 {
-		return nil
-	}
-	ops := make([]TxnOp, len(msgs))
-	for i, m := range msgs {
-		ops[i] = TxnOp{
-			Type:  TxnOpType(m.Type),
-			Key:   m.Key,
-			Value: m.Value,
-			Lease: m.Lease,
-		}
-	}
-	return ops
-}
-
-// prepareTxn evaluates all conditions and prepares the WAL entry for the
-// selected branch. Must be called under n.mu.
-//
-// Returns a zero-valued Entry (Op==0) when the branch has no write ops.
-func (n *Node) prepareTxn(req TxnRequest) (wal.Entry, bool, map[string]struct{}, error) {
-	// Evaluate conditions.
-	succeeded := true
-	for _, cond := range req.Conditions {
-		existing, err := n.readKey(cond.Key)
-		if err != nil {
-			return wal.Entry{}, false, nil, err
-		}
-		if !txnCondMatches(cond, existing) {
-			succeeded = false
-			break
-		}
-	}
-
-	ops := req.Success
-	if !succeeded {
-		ops = req.Failure
-	}
-	if len(ops) == 0 {
-		return wal.Entry{}, succeeded, nil, nil
-	}
-
-	// Pre-resolve all ops (read current state) before incrementing nextRev,
-	// so that a lookup error cannot leave nextRev in a skipped state.
-	type resolvedOp struct {
-		walOp          wal.Op
-		key            string
-		value          []byte
-		lease          int64
-		createRevision int64
-		prevRevision   int64
-		skip           bool // delete of a non-existent key
-	}
-	resolved := make([]resolvedOp, 0, len(ops))
-	for _, op := range ops {
-		existing, err := n.readKey(op.Key)
-		if err != nil {
-			return wal.Entry{}, false, nil, err
-		}
-		switch op.Type {
-		case TxnPut:
-			var walOp wal.Op
-			var createRev, prevRev int64
-			if existing == nil {
-				walOp = wal.OpCreate
-			} else {
-				walOp = wal.OpUpdate
-				createRev = existing.CreateRevision
-				prevRev = existing.Revision
-			}
-			resolved = append(resolved, resolvedOp{
-				walOp: walOp, key: op.Key, value: op.Value, lease: op.Lease,
-				createRevision: createRev, prevRevision: prevRev,
-			})
-		case TxnDelete:
-			if existing == nil {
-				resolved = append(resolved, resolvedOp{skip: true})
-				continue
-			}
-			resolved = append(resolved, resolvedOp{
-				walOp: wal.OpDelete, key: op.Key,
-				createRevision: existing.CreateRevision, prevRevision: existing.Revision,
-			})
-		}
-	}
-
-	// Drop skipped ops; if nothing remains this is effectively a no-op.
-	active := resolved[:0]
-	for _, r := range resolved {
-		if !r.skip {
-			active = append(active, r)
-		}
-	}
-	if len(active) == 0 {
-		return wal.Entry{}, succeeded, nil, nil
-	}
-
-	if len(active) > 65535 {
-		return wal.Entry{}, false, nil, fmt.Errorf("txn: too many ops (%d), maximum is 65535", len(active))
-	}
-
-	seen := make(map[string]struct{}, len(active))
-	for _, r := range active {
-		if _, dup := seen[r.key]; dup {
-			return wal.Entry{}, false, nil, fmt.Errorf("txn: duplicate key %q in branch", r.key)
-		}
-		seen[r.key] = struct{}{}
-	}
-
-	n.nextRev++
-	newRev := n.nextRev
-
-	var deletedKeys map[string]struct{}
-	subOps := make([]wal.TxnSubOp, len(active))
-	for i, r := range active {
-		cr := r.createRevision
-		if r.walOp == wal.OpCreate {
-			cr = newRev
-		}
-		subOps[i] = wal.TxnSubOp{
-			Op: r.walOp, Key: r.key, Value: r.value, Lease: r.lease,
-			CreateRevision: cr, PrevRevision: r.prevRevision,
-		}
-		if r.walOp == wal.OpDelete {
-			n.pending[r.key] = pendingKV{rev: newRev, deleted: true}
-			if deletedKeys == nil {
-				deletedKeys = make(map[string]struct{})
-			}
-			deletedKeys[r.key] = struct{}{}
-		} else {
-			n.pending[r.key] = pendingKV{
-				rev: newRev,
-				kv: &istore.KeyValue{
-					Key: r.key, Value: r.value, Revision: newRev,
-					CreateRevision: cr, PrevRevision: r.prevRevision,
-					Lease: r.lease,
-				},
-			}
-		}
-	}
-
-	return wal.Entry{
-		Revision: newRev,
-		Term:     n.term,
-		Op:       wal.OpTxn,
-		Value:    wal.EncodeTxnOps(subOps),
-	}, succeeded, deletedKeys, nil
-}
-
-func opLabel(op wal.Op) string {
-	switch op {
-	case wal.OpCreate:
-		return "create"
-	case wal.OpUpdate:
-		return "update"
-	case wal.OpDelete:
-		return "delete"
-	case wal.OpCompact:
-		return "compact"
-	case wal.OpTxn:
-		return "txn"
-	default:
-		return "unknown"
-	}
-}
-
-// await waits for a commit-loop result. req must have been sent to writeC
-// before n.mu was released. key and rev identify the pending map entry to
-// remove once the commit completes (pass key=="" for Compact which has none).
-func (n *Node) await(ctx context.Context, req *writeReq, op string, start time.Time, key string, rev int64) (int64, error) {
-	cleanPending := func() {
-		if key != "" {
-			n.mu.Lock()
-			if p, ok := n.pending[key]; ok && p.rev == rev {
-				delete(n.pending, key)
-			}
-			n.mu.Unlock()
-		}
-	}
-	var err error
-	select {
-	case err = <-req.done:
-	case <-ctx.Done():
-		// Give the commit loop a brief window to react: it may have already
-		// detected our cancellation (via batchCtx) and is about to signal us
-		// with a meaningful error. If it does, use that result instead of
-		// ctx.Err() so callers see a system-level error rather than their own
-		// deadline expiry.
-		timer := time.NewTimer(5 * time.Millisecond)
-		defer timer.Stop()
-		select {
-		case err = <-req.done:
-			// commit loop responded promptly; fall through to normal handling
-		case <-timer.C:
-			go func() { <-req.done; cleanPending() }()
-			return 0, ctx.Err()
-		}
-	}
-	cleanPending()
-	if err != nil {
-		metrics.WriteErrors.WithLabelValues(op).Inc()
-		return 0, fmt.Errorf("t4: commit: %w", err)
-	}
-	metrics.WritesTotal.WithLabelValues(op).Inc()
-	metrics.WriteDuration.WithLabelValues(op).Observe(time.Since(start).Seconds())
-	metrics.CurrentRevision.Set(float64(rev))
-	count := atomic.AddInt64(&n.entriesSinceCheckpoint, 1)
-	if n.checkpointTriggerC != nil && count >= n.cfg.CheckpointEntries {
-		select {
-		case n.checkpointTriggerC <- struct{}{}:
-		default: // already a pending trigger; don't block
-		}
-	}
-	return rev, nil
-}
-
-// clearPendingBatch removes optimistic pending entries for one commit-loop
-// batch. If a newer write has already reused the same key, the revision guard
-// preserves that newer pending entry.
-func (n *Node) clearPendingBatch(batch []*writeReq) {
-	n.mu.Lock()
-	defer n.mu.Unlock()
-	for _, req := range batch {
-		if req.entry.Op == wal.OpTxn {
-			// For txn entries the key field is empty; decode sub-ops to clear
-			// each affected key from the pending map.
-			if ops, err := wal.DecodeTxnOps(req.entry.Value); err == nil {
-				for _, op := range ops {
-					if p, ok := n.pending[op.Key]; ok && p.rev == req.entry.Revision {
-						delete(n.pending, op.Key)
-					}
-				}
-			}
-			continue
-		}
-		if req.entry.Key == "" {
-			continue
-		}
-		if p, ok := n.pending[req.entry.Key]; ok && p.rev == req.entry.Revision {
-			delete(n.pending, req.entry.Key)
-		}
-	}
-}
-
-// commitLoop is the group-commit pipeline for leader/single-node writes.
-// It drains writeC, writes all entries to WAL with a single fsync, applies
-// them to Pebble as a batch, and signals each caller's done channel.
-func (n *Node) commitLoop(ctx context.Context) {
-	defer func() {
-		// Fence the node first so new writers fail fast.
-		n.closed.Store(true)
-
-		// Drain requests immediately to free queue slots. This unblocks writers
-		// that might be stuck on n.writeC <- req while holding n.mu.
-	drain:
-		for {
-			select {
-			case req := <-n.writeC:
-				req.done <- ErrClosed
-			default:
-				break drain
-			}
-		}
-
-		// Wait for any writer currently in the critical section (between closed
-		// check and queue send) to finish, then perform a final drain pass.
-		n.mu.Lock()
-		n.mu.Unlock()
-		for {
-			select {
-			case req := <-n.writeC:
-				req.done <- ErrClosed
-			default:
-				return
-			}
-		}
-	}()
-
-	for {
-		// Block until at least one request arrives.
-		var batch []*writeReq
-		select {
-		case req := <-n.writeC:
-			batch = append(batch, req)
-		case <-ctx.Done():
-			return
-		}
-		// Drain any additional requests that arrived while we were processing.
-	drain:
-		for {
-			select {
-			case req := <-n.writeC:
-				batch = append(batch, req)
-			default:
-				break drain
-			}
-		}
-
-		// Build a context that's cancelled when any batch caller gives up.
-		// This lets the WAL abort early (in tests / context-aware WALs) if
-		// all callers have abandoned the batch. Real fsyncs complete normally.
-		batchCtx, batchCancel := context.WithCancel(ctx)
-		for _, req := range batch {
-			r := req
-			go func() {
-				select {
-				case <-r.ctx.Done():
-					batchCancel()
-				case <-batchCtx.Done():
-				}
-			}()
-		}
-
-		// Write all entries to WAL with one fsync.
-		entries := make([]*wal.Entry, len(batch))
-		for i, req := range batch {
-			entries[i] = &req.entry
-		}
-
-		var err error
-		if n.peerSrv != nil {
-			// Pipeline: overlap network delivery to followers with the leader's
-			// own WAL fsync, but do not let followers make entries durable or
-			// visible until the leader has fsynced successfully. Followers stage
-			// entry messages in memory and wait for BroadcastCommit before
-			// appending/applying the batch locally.
-			walErrC := make(chan error, 1)
-			go func() { walErrC <- n.wal.AppendBatch(batchCtx, entries) }()
-
-			for _, req := range batch {
-				n.peerSrv.Broadcast(&req.entry)
-			}
-
-			startRev := batch[0].entry.Revision
-			maxRev := batch[len(batch)-1].entry.Revision
-			err = <-walErrC
-			if err == nil {
-				n.peerSrv.BroadcastCommit(startRev, maxRev)
-
-				// Wait for follower ACKs according to the configured policy.
-				// Use the commit loop's own context (node lifetime), NOT batchCtx:
-				// batchCtx is cancelled after AppendBatch returns and passing it
-				// here would cause WaitForFollowers to return instantly.
-				//
-				// Availability policy: if all followers disconnect mid-wait, we
-				// proceed anyway — the entry is already durable in the leader's
-				// WAL and will be replayed by followers when they reconnect.
-				if waitErr := n.peerSrv.WaitForFollowers(ctx, maxRev, peer.WaitMode(n.cfg.FollowerWaitMode)); waitErr != nil {
-					err = waitErr
-				}
-			}
-		} else {
-			err = n.wal.AppendBatch(batchCtx, entries)
-		}
-		batchCancel() // release watcher goroutines
-
-		// Apply all entries to Pebble as one batch (in order).
-		if err == nil {
-			dbEntries := make([]wal.Entry, len(batch))
-			for i, req := range batch {
-				dbEntries[i] = req.entry
-			}
-			err = n.db.Load().Apply(dbEntries)
-		}
-
-		// Clear optimistic state before waking callers so a failed batch cannot
-		// leak stale pending revisions into a racing follow-up write.
-		n.clearPendingBatch(batch)
-
-		// Signal all callers.
-		for _, req := range batch {
-			req.done <- err
-		}
-		if err != nil {
-			if errors.Is(err, context.Canceled) || errors.Is(err, context.DeadlineExceeded) {
-				// Callers abandoned the batch; this is not a permanent fault.
-				// Let the loop continue for the next batch.
-				continue
-			}
-			// A WAL or Pebble error leaves the segment in an unknown state.
-			// Stop accepting writes immediately; the defer will fence the node.
-			return
-		}
-	}
-}
-
-// Compact removes log entries at or below revision.
-func (n *Node) Compact(ctx context.Context, revision int64) error {
-	if n.loadRole() == roleFollower {
-		resp, err := n.forwardWrite(ctx, &peer.ForwardRequest{Op: peer.ForwardCompact, Revision: revision})
-		if err != nil {
-			return err
-		}
-		return decodeErr(resp.ErrCode, resp.ErrMsg)
-	}
-	n.fenceMu.RLock()
-	defer n.fenceMu.RUnlock()
-	start := time.Now()
-	n.mu.Lock()
-	if n.closed.Load() {
-		n.mu.Unlock()
-		return ErrClosed
-	}
-	n.nextRev++
-	e := wal.Entry{
-		Revision: n.nextRev, Term: n.term, Op: wal.OpCompact,
-		PrevRevision: revision,
-	}
-	req := newWriteReq(ctx, e)
-	n.writeC <- req
-	n.mu.Unlock()
-	_, err := n.await(ctx, req, "compact", start, "", e.Revision)
-	return err
-}
-
 // ── Read path (all roles serve locally) ──────────────────────────────────────
+
+// ReadConsistency returns the configured read consistency mode.
+func (n *Node) ReadConsistency() ReadConsistency { return n.cfg.ReadConsistency }
+
+// syncWithLeader implements the ReadIndex pattern: ask the leader for its
+// current revision, then wait until this node has applied at least that far.
+// Returns nil immediately if the node is the leader or running single-node.
+func (n *Node) syncWithLeader(ctx context.Context) error {
+	cli := n.leaderCli.Load()
+	if cli == nil {
+		// If the background context has been cancelled the node is either
+		// shutting down or has been fenced (leader superseded by a new term).
+		// Serving a read from our local stale Pebble would violate
+		// linearizability — return an error so the client retries elsewhere.
+		if n.bgCtx.Err() != nil {
+			return ErrClosed
+		}
+		return nil // leader or single-node — already up-to-date
+	}
+	resp, err := cli.ForwardWrite(ctx, &peer.ForwardRequest{Op: peer.ForwardGetRevision})
+	if err != nil {
+		return fmt.Errorf("t4: read sync: %w", err)
+	}
+	if err := n.WaitForRevision(ctx, resp.Revision); err != nil {
+		if errors.Is(err, context.DeadlineExceeded) || errors.Is(err, context.Canceled) {
+			return fmt.Errorf("t4: read sync: leader reported rev=%d but local node only reached rev=%d before wait ended: %w",
+				resp.Revision, n.db.Load().CurrentRevision(), err)
+		}
+		return fmt.Errorf("t4: read sync: wait for local revision %d: %w", resp.Revision, err)
+	}
+	return nil
+}
+
+// LinearizableGet returns the value for key with linearizability guaranteed.
+// On a follower it syncs to the leader's revision before serving locally.
+func (n *Node) LinearizableGet(ctx context.Context, key string) (*KeyValue, error) {
+	if err := n.syncWithLeader(ctx); err != nil {
+		return nil, err
+	}
+	return n.Get(key)
+}
+
+// LinearizableList returns all keys with the given prefix with linearizability guaranteed.
+func (n *Node) LinearizableList(ctx context.Context, prefix string) ([]*KeyValue, error) {
+	if err := n.syncWithLeader(ctx); err != nil {
+		return nil, err
+	}
+	return n.List(prefix)
+}
+
+// LinearizableCount returns the count of keys with the given prefix with linearizability guaranteed.
+func (n *Node) LinearizableCount(ctx context.Context, prefix string) (int64, error) {
+	if err := n.syncWithLeader(ctx); err != nil {
+		return 0, err
+	}
+	return n.Count(prefix)
+}
 
 func (n *Node) Get(key string) (*KeyValue, error) {
 	if n.closed.Load() {
@@ -2379,186 +870,6 @@ func (n *Node) Watch(ctx context.Context, prefix string, startRev int64) (<-chan
 	return out, nil
 }
 
-// uploadLocalWALSegments uploads any sealed local WAL segment files that are
-// not yet present in S3. This is called when becoming leader so that local
-// entries (recovered via replayLocal) are durable in object storage before
-// followers can bootstrap.
-func uploadLocalWALSegments(ctx context.Context, walDir string, store object.Store, log Logger) {
-	paths, err := wal.LocalSegments(walDir)
-	if err != nil || len(paths) == 0 {
-		return
-	}
-
-	// Build set of keys already in S3 to skip redundant uploads.
-	s3Keys, _ := store.List(ctx, "wal/")
-	inS3 := make(map[string]struct{}, len(s3Keys))
-	for _, k := range s3Keys {
-		inS3[k] = struct{}{}
-	}
-
-	up := makeUploader(store, log)
-	for _, path := range paths {
-		term, firstRev, ok := wal.ParseSegmentName(filepath.Base(path))
-		if !ok {
-			continue
-		}
-		objKey := wal.ObjectKey(term, firstRev)
-		if _, exists := inS3[objKey]; exists {
-			continue // already uploaded
-		}
-		if err := up(ctx, path, objKey); err != nil {
-			log.Warnf("t4: pre-leader upload %q → %q: %v", path, objKey, err)
-		}
-	}
-}
-
-// ── Background checkpoint loop ────────────────────────────────────────────────
-
-func (n *Node) checkpointLoop(ctx context.Context) {
-	// Write an immediate checkpoint before entering the ticker so that any
-	// entries recovered from local WAL segments (but not yet in S3) are
-	// captured in the checkpoint. Without this, a crash after becoming leader
-	// but before the first periodic checkpoint could leave new followers unable
-	// to see those entries.
-	n.forceCheckpoint(ctx)
-
-	ticker := time.NewTicker(n.cfg.CheckpointInterval)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-ticker.C:
-			n.maybeCheckpoint(ctx)
-		case <-n.checkpointTriggerC:
-			n.maybeCheckpoint(ctx)
-		case <-ctx.Done():
-			return
-		}
-	}
-}
-
-// forceCheckpoint writes a checkpoint unconditionally (bypassing the
-// entriesSinceCheckpoint guard). Used on startup to capture local state.
-func (n *Node) forceCheckpoint(ctx context.Context) {
-	n.fenceMu.Lock()
-	rev := n.db.Load().CurrentRevision()
-	if rev == 0 {
-		n.fenceMu.Unlock()
-		return
-	}
-	if err := n.wal.SealAndFlush(rev + 1); err != nil {
-		n.fenceMu.Unlock()
-		n.log.Errorf("t4: startup checkpoint seal WAL: %v", err)
-		return
-	}
-	if err := n.db.Load().Flush(); err != nil {
-		n.fenceMu.Unlock()
-		n.log.Errorf("t4: startup checkpoint flush pebble: %v", err)
-		return
-	}
-	if n.sstUploader != nil {
-		n.sstUploader.Wait()
-		if err := n.cp.WriteWithRegistry(ctx, n.db.Load().Pebble(), n.cfg.ObjectStore, n.term, rev, "", n.sstUploader.Registry(), n.sstUploader.InheritedRegistry()); err != nil {
-			n.fenceMu.Unlock()
-			n.log.Errorf("t4: startup checkpoint rev=%d: %v", rev, err)
-			return
-		}
-	} else if err := n.cp.Write(ctx, n.db.Load().Pebble(), n.cfg.ObjectStore, n.term, rev, "", n.cfg.AncestorStore); err != nil {
-		n.fenceMu.Unlock()
-		n.log.Errorf("t4: startup checkpoint rev=%d: %v", rev, err)
-		return
-	}
-	n.fenceMu.Unlock()
-	atomic.StoreInt64(&n.entriesSinceCheckpoint, 0)
-	metrics.CheckpointsTotal.Inc()
-	n.log.Infof("t4: startup checkpoint written (rev=%d)", rev)
-}
-
-func (n *Node) maybeCheckpoint(ctx context.Context) {
-	if atomic.LoadInt64(&n.entriesSinceCheckpoint) == 0 {
-		return
-	}
-	n.fenceMu.Lock()
-	rev := n.db.Load().CurrentRevision()
-	if rev == 0 {
-		n.fenceMu.Unlock()
-		return
-	}
-	if err := n.wal.SealAndFlush(rev + 1); err != nil {
-		n.fenceMu.Unlock()
-		n.log.Errorf("t4: checkpoint seal WAL: %v", err)
-		return
-	}
-	if err := n.db.Load().Flush(); err != nil {
-		n.fenceMu.Unlock()
-		n.log.Errorf("t4: checkpoint flush pebble: %v", err)
-		return
-	}
-	if n.sstUploader != nil {
-		n.sstUploader.Wait()
-		if err := n.cp.WriteWithRegistry(ctx, n.db.Load().Pebble(), n.cfg.ObjectStore, n.term, rev, "", n.sstUploader.Registry(), n.sstUploader.InheritedRegistry()); err != nil {
-			n.fenceMu.Unlock()
-			n.log.Errorf("t4: write checkpoint rev=%d: %v", rev, err)
-			return
-		}
-	} else if err := n.cp.Write(ctx, n.db.Load().Pebble(), n.cfg.ObjectStore, n.term, rev, "", n.cfg.AncestorStore); err != nil {
-		n.fenceMu.Unlock()
-		n.log.Errorf("t4: write checkpoint rev=%d: %v", rev, err)
-		return
-	}
-	n.fenceMu.Unlock()
-	atomic.StoreInt64(&n.entriesSinceCheckpoint, 0)
-	metrics.CheckpointsTotal.Inc()
-	n.log.Infof("t4: checkpoint written (rev=%d)", rev)
-
-	// GC WAL segments from S3 that are fully covered by this checkpoint AND
-	// that all connected followers have applied. Using min(leaderRev,
-	// minFollowerAppliedRev) as the GC boundary ensures we never delete a
-	// segment that a follower still needs to replay.
-	gcCtx, cancel := context.WithTimeout(ctx, 2*time.Minute)
-	defer cancel()
-	gcRev := rev
-	if n.peerSrv != nil {
-		if minFollower := n.peerSrv.MinFollowerAppliedRev(); minFollower < gcRev {
-			gcRev = minFollower
-		}
-	}
-	deleted, gcErr := wal.GCSegments(gcCtx, n.cfg.ObjectStore, gcRev, n.log)
-	if gcErr != nil {
-		n.log.Warnf("t4: wal gc: %v", gcErr)
-	} else if deleted > 0 {
-		metrics.WALGCTotal.Add(float64(deleted))
-		n.log.Infof("t4: wal gc: deleted %d segments (covered by checkpoint rev=%d)", deleted, gcRev)
-	}
-
-	// GC old checkpoint archives from S3, keeping the 2 most recent so that
-	// any in-flight bootstrap that read manifest/latest just before we
-	// overwrote it can still fetch the previous checkpoint.
-	// GCCheckpoints deletes old checkpoint archives and returns the set of SST
-	// keys that were exclusively referenced by the deleted checkpoints. Passing
-	// that candidate set to GCOrphanSSTs (instead of listing all "sst/" keys)
-	// eliminates the race where a newly-promoted leader uploads SSTs before
-	// writing its first checkpoint — those SSTs never appear in any deleted
-	// checkpoint's index, so they are never mistakenly treated as orphans.
-	cpDeleted, orphanSSTs, cpGCErr := n.cp.GCCheckpoints(gcCtx, n.cfg.ObjectStore, 2)
-	if cpGCErr != nil {
-		n.log.Warnf("t4: checkpoint gc: %v", cpGCErr)
-	} else if cpDeleted > 0 {
-		n.log.Infof("t4: checkpoint gc: deleted %d old checkpoint(s)", cpDeleted)
-	}
-
-	// Only run SST GC when old checkpoints were actually deleted and there are
-	// candidate SSTs to clean up. This skips the Delete loop entirely when
-	// nothing changed, which is the common case.
-	if cpDeleted > 0 && len(orphanSSTs) > 0 {
-		sstDeleted, sstGCErr := n.cp.GCOrphanSSTs(gcCtx, n.cfg.ObjectStore, orphanSSTs)
-		if sstGCErr != nil {
-			n.log.Warnf("t4: sst gc: %v", sstGCErr)
-		} else if sstDeleted > 0 {
-			n.log.Infof("t4: sst gc: deleted %d orphan sst(s)", sstDeleted)
-		}
-	}
-}
-
 // ── Helpers ───────────────────────────────────────────────────────────────────
 
 func toKV(sv *istore.KeyValue) *KeyValue {
@@ -2605,345 +916,4 @@ func makeUploader(obj object.Store, log Logger) wal.Uploader {
 		metrics.WALUploadDuration.Observe(time.Since(start).Seconds())
 		return os.Remove(localPath)
 	}
-}
-
-func replayLocal(db *istore.Store, walDir string, afterRev int64, log Logger) error {
-	paths, err := wal.LocalSegments(walDir)
-	if err != nil {
-		return err
-	}
-	for _, path := range paths {
-		sr, closer, err := wal.OpenSegmentFile(path)
-		if err != nil {
-			return err
-		}
-		entries, readErr := sr.ReadAll()
-		closer()
-		if readErr != nil {
-			log.Warnf("t4: partial local segment %q: %v", path, readErr)
-		}
-		var applicable []wal.Entry
-		for _, e := range entries {
-			if e.Revision > afterRev {
-				applicable = append(applicable, *e)
-			}
-		}
-		if len(applicable) > 0 {
-			if err := db.Recover(applicable); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// replayPinned replays the specific WAL segments listed in rp, applying
-// entries with revision > afterRev. Used during RestorePoint bootstrap.
-func replayPinned(ctx context.Context, db *istore.Store, rp *RestorePoint, afterRev int64, log Logger) error {
-	for _, seg := range rp.WALSegments {
-		rc, err := rp.Store.GetVersioned(ctx, seg.Key, seg.VersionID)
-		if err != nil {
-			return fmt.Errorf("replayPinned get %q@%s: %w", seg.Key, seg.VersionID, err)
-		}
-		sr, err := wal.NewSegmentReader(rc)
-		if err != nil {
-			rc.Close()
-			return fmt.Errorf("replayPinned segment %q: %w", seg.Key, err)
-		}
-		entries, readErr := sr.ReadAll()
-		rc.Close()
-		if readErr != nil {
-			log.Warnf("t4: partial pinned segment %q: %v", seg.Key, readErr)
-		}
-		var applicable []wal.Entry
-		for _, e := range entries {
-			if e.Revision > afterRev {
-				applicable = append(applicable, *e)
-			}
-		}
-		if len(applicable) > 0 {
-			if err := db.Recover(applicable); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-// restoreDBIfBehindCheckpoint checks whether the node's Pebble database is
-// behind the latest S3 checkpoint. If so, it performs an in-place restore:
-//
-//  1. (No lock) Download checkpoint SSTs to a temp directory, retrying if the
-//     checkpoint was GC'd between the manifest read and the download.
-//  2. (fenceMu.Lock) Close the old Pebble, atomic-rename the temp dir to the
-//     canonical db/ path, open the new Pebble. fenceMu blocks all concurrent
-//     read/write handlers for the brief swap window.
-//
-// On success n.db and n.term are updated and the function returns true.
-// Returns (false, nil) when no restore was needed (already at or above
-// the latest checkpoint revision).
-func (n *Node) restoreDBIfBehindCheckpoint(ctx context.Context) (bool, error) {
-	manifest, err := n.cp.ReadManifest(ctx, n.cfg.ObjectStore)
-	if err != nil {
-		return false, fmt.Errorf("read manifest: %w", err)
-	}
-	if manifest == nil || manifest.Revision <= n.db.Load().CurrentRevision() {
-		return false, nil // already at or above checkpoint
-	}
-	n.log.Infof("t4: node at rev=%d is behind checkpoint rev=%d — restoring in place",
-		n.db.Load().CurrentRevision(), manifest.Revision)
-
-	pebbleDir := filepath.Join(n.cfg.DataDir, "db")
-	tmpDir := pebbleDir + ".resync"
-
-	// ── Phase 1: restore checkpoint SSTs to a temp dir (no lock held) ───────
-	var newTerm uint64
-	for attempt := 0; ; attempt++ {
-		os.RemoveAll(tmpDir)
-		t, _, rerr := n.cp.Restore(ctx, n.cfg.ObjectStore, manifest.CheckpointKey, tmpDir)
-		if rerr == nil {
-			newTerm = t
-			break
-		}
-		if !errors.Is(rerr, object.ErrNotFound) || attempt >= 4 {
-			os.RemoveAll(tmpDir)
-			return false, fmt.Errorf("restore checkpoint: %w", rerr)
-		}
-		// Checkpoint was GC'd between manifest read and restore — re-read
-		// the manifest so we target whatever the leader most recently wrote.
-		n.log.Warnf("t4: checkpoint %q not found during restore, re-reading manifest (attempt %d/5)",
-			manifest.CheckpointKey, attempt+1)
-		time.Sleep(500 * time.Millisecond)
-		manifest, err = n.cp.ReadManifest(ctx, n.cfg.ObjectStore)
-		if err != nil {
-			os.RemoveAll(tmpDir)
-			return false, fmt.Errorf("re-read manifest: %w", err)
-		}
-		if manifest == nil {
-			os.RemoveAll(tmpDir)
-			return false, fmt.Errorf("manifest disappeared during restore")
-		}
-	}
-
-	// ── Phase 2: swap Pebble directories under fenceMu.Lock + readMu.Lock ───
-	// fenceMu.Lock drains in-flight writes (which hold fenceMu.RLock).
-	//
-	// We first signal-close the old store before acquiring readMu.Lock.
-	// SignalClose unblocks any goroutine blocked in WaitForRevision
-	// (those goroutines hold readMu.RLock while waiting). If we acquired
-	// readMu.Lock first, WaitForRevision callers would never release their
-	// RLocks → deadlock.
-	//
-	// After SignalClose, readMu.Lock drains the remaining in-flight reads
-	// (Get/List and any WaitForRevision callers that were unblocked by
-	// SignalClose). We then close the old store while still holding readMu,
-	// so no new read can load the old *Store after Pebble has been closed.
-	// We keep readMu until n.db.Store(newDB) so no read can observe the
-	// closed store pointer during the on-disk swap.
-	//
-	// Lock order: fenceMu → readMu (no other code path acquires them in the
-	// reverse order, so no deadlock is possible here).
-	n.fenceMu.Lock()
-	oldDB := n.db.Load()
-	oldDB.SignalClose() // unblocks WaitForRevision waiters before we take readMu.Lock
-	n.readMu.Lock()
-	if rerr := oldDB.Close(); rerr != nil {
-		n.readMu.Unlock()
-		n.fenceMu.Unlock()
-		os.RemoveAll(tmpDir)
-		return false, fmt.Errorf("close old pebble before restore swap: %w", rerr)
-	}
-	if rerr := os.RemoveAll(pebbleDir); rerr != nil {
-		n.readMu.Unlock()
-		n.fenceMu.Unlock()
-		os.RemoveAll(tmpDir)
-		return false, fmt.Errorf("remove old pebble dir: %w", rerr)
-	}
-	if rerr := os.Rename(tmpDir, pebbleDir); rerr != nil {
-		n.readMu.Unlock()
-		n.fenceMu.Unlock()
-		os.RemoveAll(tmpDir)
-		return false, fmt.Errorf("rename resync dir: %w", rerr)
-	}
-	newDB, rerr := istore.Open(pebbleDir, n.log)
-	if rerr != nil {
-		n.readMu.Unlock()
-		n.fenceMu.Unlock()
-		return false, fmt.Errorf("open new pebble after restore: %w", rerr)
-	}
-	n.db.Store(newDB)
-	n.readMu.Unlock()
-	n.term = newTerm
-	n.fenceMu.Unlock()
-	return true, nil
-}
-
-// resyncFromCheckpoint is called from followLoop when IsResyncRequired fires.
-// It uses restoreDBIfBehindCheckpoint to close the WAL gap, then (if a restore
-// was actually performed) replaces the local WAL so subsequent Appends from the
-// live stream start at the correct revision.
-func (n *Node) resyncFromCheckpoint(bgCtx context.Context) error {
-	ctx, cancel := context.WithTimeout(bgCtx, 5*time.Minute)
-	defer cancel()
-
-	restored, err := n.restoreDBIfBehindCheckpoint(ctx)
-	if err != nil {
-		return err
-	}
-	if !restored {
-		return nil
-	}
-
-	// ── Phase 3: replace WAL and update node metadata ────────────────────────
-	// followLoop is the sole WAL writer for a follower, so no concurrent
-	// Append calls can race with this replacement.
-	walDir := filepath.Join(n.cfg.DataDir, "wal")
-	newRev := n.db.Load().CurrentRevision()
-	n.wal.Close()
-	if rerr := os.RemoveAll(walDir); rerr != nil {
-		n.log.Warnf("t4: remove old wal dir during resync: %v", rerr)
-	}
-	newWal, rerr := wal.Open(walDir, n.term, newRev+1,
-		wal.WithSegmentMaxSize(n.cfg.SegmentMaxSize),
-		wal.WithSegmentMaxAge(n.cfg.SegmentMaxAge),
-		wal.WithLogger(n.log),
-	)
-	if rerr != nil {
-		return fmt.Errorf("open new wal after resync: %w", rerr)
-	}
-	newWal.Start(bgCtx)
-
-	n.mu.Lock()
-	n.wal = newWal
-	n.nextRev = newRev
-	n.mu.Unlock()
-
-	n.log.Infof("t4: follower in-place resync complete (rev=%d term=%d)", newRev, n.term)
-	return nil
-}
-
-func replayRemote(ctx context.Context, db *istore.Store, obj object.Store, afterRev int64, log Logger) error {
-	keys, err := obj.List(ctx, "wal/")
-	if err != nil {
-		return err
-	}
-
-	// Build per-term cutoffs to handle leader-change conflicts.
-	//
-	// When a new leader takes over at revision X, it starts writing a new term
-	// from revision X+1. The old leader may have written S3 segments that cover
-	// some of those same revisions (X+1, X+2, …). Applying both the old-term
-	// and new-term entries at the same revision would corrupt the index.
-	//
-	// cutoff[term] is the minimum firstRev among all segments with term > term.
-	// Old-term entries at revision >= cutoff are superseded by the new term and
-	// must be skipped. For the highest term present, cutoff = math.MaxInt64 (no
-	// upper bound — apply all entries).
-	cutoff := walTermCutoffs(keys)
-
-	var all []wal.Entry
-	for _, key := range keys {
-		term, _ := parseWALKey(key)
-		termCutoff := cutoff[term]
-
-		rc, err := obj.Get(ctx, key)
-		if err != nil {
-			return fmt.Errorf("replayRemote get %q: %w", key, err)
-		}
-		sr, err := wal.NewSegmentReader(rc)
-		if err != nil {
-			rc.Close()
-			return fmt.Errorf("replayRemote segment %q: %w", key, err)
-		}
-		entries, readErr := sr.ReadAll()
-		rc.Close()
-		if readErr != nil {
-			log.Warnf("t4: partial remote segment %q: %v", key, readErr)
-		}
-		for _, e := range entries {
-			if e.Revision <= afterRev {
-				continue // already covered by checkpoint / local WAL
-			}
-			if e.Revision >= termCutoff {
-				continue // superseded by a higher-term entry at this revision
-			}
-			all = append(all, *e)
-		}
-	}
-
-	if len(all) == 0 {
-		return nil
-	}
-
-	// Ensure deterministic order and resolve same-revision duplicates by
-	// keeping the highest-term entry at each revision.
-	sort.Slice(all, func(i, j int) bool {
-		if all[i].Revision != all[j].Revision {
-			return all[i].Revision < all[j].Revision
-		}
-		return all[i].Term < all[j].Term
-	})
-	merged := make([]wal.Entry, 0, len(all))
-	for _, e := range all {
-		if n := len(merged); n > 0 && merged[n-1].Revision == e.Revision {
-			if e.Term >= merged[n-1].Term {
-				merged[n-1] = e
-			}
-			continue
-		}
-		merged = append(merged, e)
-	}
-
-	// Fail closed on holes: a missing revision means we likely raced WAL GC
-	// during bootstrap/resync and must restore from a fresher checkpoint.
-	expected := afterRev + 1
-	for _, e := range merged {
-		if e.Revision != expected {
-			return fmt.Errorf("replayRemote: missing revision(s): expected %d got %d", expected, e.Revision)
-		}
-		expected++
-	}
-
-	if err := db.Recover(merged); err != nil {
-		return err
-	}
-	return nil
-}
-
-// walTermCutoffs returns a map from term → the minimum firstRev of any
-// segment whose term is strictly greater. Entries from a given term at
-// revisions >= cutoff are superseded by the newer term and should be skipped.
-// The highest term maps to math.MaxInt64 (no cutoff).
-func walTermCutoffs(keys []string) map[uint64]int64 {
-	// Collect the minimum firstRev for each term.
-	minFirstRev := map[uint64]int64{}
-	for _, key := range keys {
-		term, firstRev := parseWALKey(key)
-		if existing, ok := minFirstRev[term]; !ok || firstRev < existing {
-			minFirstRev[term] = firstRev
-		}
-	}
-	cutoff := make(map[uint64]int64, len(minFirstRev))
-	for term := range minFirstRev {
-		c := int64(math.MaxInt64)
-		for t, fr := range minFirstRev {
-			if t > term && fr < c {
-				c = fr
-			}
-		}
-		cutoff[term] = c
-	}
-	return cutoff
-}
-
-// parseWALKey extracts term and firstRev from a WAL object key of the form
-// "wal/{term:010d}/{firstRev:020d}".
-func parseWALKey(key string) (term uint64, firstRev int64) {
-	parts := strings.SplitN(key, "/", 3)
-	if len(parts) != 3 {
-		return 0, 0
-	}
-	term64, _ := strconv.ParseUint(strings.TrimLeft(parts[1], "0"), 10, 64)
-	rev, _ := strconv.ParseInt(strings.TrimLeft(parts[2], "0"), 10, 64)
-	return term64, rev
 }
